Fast fertig
git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C1284

+++ src/Game/SceneAdrian.cpp
@@ -76,7 +76,7 @@ namespace TikiEngine
 				IAnimation* anim = go->GModel()->GetAnimation(0);
 				anim->SetAnimationSpeed(5.0);
 				go->PRS.SScale() = Vector3(0.01f);
-				go->PRS.SPosition() = Vector3(i , 0, 0);
+				go->PRS.SPosition() = Vector3((float)i , 0, 0);
 				go->GetSceneGraphElement().SetDynamic();
 
 				this->AddElement(go);
